requirements.txt
# Phase 1 dependencies
requests
python-dotenv